Move redis operations to concern
